refactor: simplify main package by reusing vx-cli main function
- Export main function from vx-cli lib to avoid code duplication
- Simplify main package dependencies by removing direct tool dependencies
- Main package now only depends on vx-cli and delegates all functionality
- Reduces maintenance cost and ensures consistency
- Fixes binary installation issue for 'cargo install vx'
